Client middleware: fixes and transpiling
