+++ client/getServerDispatchMidddleware.js
@@ -1,44 +1,1 @@
-'use strict';
-
-var isPlainObject = require('node_modules/lodash/isPlainObject');
-var assign = require('node_modules/lodash/assign');
-
-var serverRegexp = /^@@server\/.*?$/i;
-var defaultOptions = {
-  next: false,
-  exclude: null,
-  include: null
-};
-
-/**
- * Process client -> server dispatching
- * @param  {Socket} io - socket.io connection to server
- * @param  {Object} options - action processing options
- * @param  {Boolean} options.next - process server action or not
- * @param  {Array} options.exclude - list of disallowed server action types
- * @param  {Array} options.include - list of allowed server action types
- * @return {Mided}  - do not use it
- */
-
-module.exports = function getServerDispatchMiddleware(io, options) {
-  var middlewareOptions = options ? assign(defaultOptions, options) : defaultOptions;
-
-  return function () {
-    return function (next) {
-      return function (action) {
-        if (!isPlainObject(action)) return next(action);
-        if (!action.type) return next(action);
-        if (serverRegexp.test(action.type)) {
-          var allowedType = (!middlewareOptions.exclude || middlewareOptions.exclude.indexOf(action.type) === -1) && (!middlewareOptions.include || middlewareOptions.include.indexOf(action.type) !== -1);
-
-          if (!middlewareOptions.next || middlewareOptions.next && !allowedType) {
-            return io.emit('dispatch', action);
-          } else {
-            io.emit('dispatch', action);
-          }
-        }
-        next(action);
-      };
-    };
-  };
-};
+!function(t){function n(e){if(o[e])return o[e].exports;var r=o[e]={i:e,l:!1,exports:{}};return t[e].call(r.exports,r,r.exports,n),r.l=!0,r.exports}var o={};n.m=t,n.c=o,n.i=function(t){return t},n.d=function(t,o,e){n.o(t,o)||Object.defineProperty(t,o,{configurable:!1,enumerable:!0,get:e})},n.n=function(t){var o=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(o,"a",o),o},n.o=function(t,n){return Object.prototype.hasOwnProperty.call(t,n)},n.p="",n(n.s=2)}([function(t,n,o){"use strict";var e=o(9),r=e.Symbol;t.exports=r},function(t,n,o){"use strict";function e(t){if(!c(t)||r(t)!=i)return!1;var n=u(t);if(null===n)return!0;var o=y.call(n,"constructor")&&n.constructor;return"function"==typeof o&&o instanceof o&&s.call(o)==p}var r=o(3),u=o(5),c=o(10),i="[object Object]",f=Function.prototype,l=Object.prototype,s=f.toString,y=l.hasOwnProperty,p=s.call(Object);t.exports=e},function(t,n,o){"use strict";var e=o(1),r=/^@@server\/.*?$/i,u={next:!1,exclude:null,include:null};t.exports=function(t,n){var o=n?Object.assign(u,n):u;return function(){return function(n){return function(u){if(!e(u))return n(u);if(!u.type)return n(u);if(r.test(u.type)){var c=!(o.exclude&&-1!==o.exclude.indexOf(u.type)||o.include&&-1===o.include.indexOf(u.type));if(!o.next||o.next&&!c)return t.emit("dispatch",u);t.emit("dispatch",u)}n(u)}}}}},function(t,n,o){"use strict";function e(t){return null==t?void 0===t?f:i:l&&l in Object(t)?u(t):c(t)}var r=o(0),u=o(6),c=o(7),i="[object Null]",f="[object Undefined]",l=r?r.toStringTag:void 0;t.exports=e},function(t,n,o){"use strict";(function(n){var o="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t},e="object"==(void 0===n?"undefined":o(n))&&n&&n.Object===Object&&n;t.exports=e}).call(n,o(11))},function(t,n,o){"use strict";var e=o(8),r=e(Object.getPrototypeOf,Object);t.exports=r},function(t,n,o){"use strict";function e(t){var n=c.call(t,f),o=t[f];try{t[f]=void 0}catch(t){}var e=i.call(t);return n?t[f]=o:delete t[f],e}var r=o(0),u=Object.prototype,c=u.hasOwnProperty,i=u.toString,f=r?r.toStringTag:void 0;t.exports=e},function(t,n,o){"use strict";function e(t){return u.call(t)}var r=Object.prototype,u=r.toString;t.exports=e},function(t,n,o){"use strict";function e(t,n){return function(o){return t(n(o))}}t.exports=e},function(t,n,o){"use strict";var e="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t},r=o(4),u="object"==("undefined"==typeof self?"undefined":e(self))&&self&&self.Object===Object&&self,c=r||u||Function("return this")();t.exports=c},function(t,n,o){"use strict";function e(t){return null!=t&&"object"==(void 0===t?"undefined":r(t))}var r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t};t.exports=e},function(t,n,o){"use strict";var e,r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t};e=function(){return this}();try{e=e||Function("return this")()||(0,eval)("this")}catch(t){"object"===("undefined"==typeof window?"undefined":r(window))&&(e=window)}t.exports=e}]);
