+++ package.json
@@ -9,7 +9,8 @@
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "eslint": "./node_modules/eslint/bin/eslint.js ./"
+    "eslint": "./node_modules/eslint/bin/eslint.js ./",
+    "build": "webpack"
   },
   "pre-commit": [
     "eslint"
@@ -22,7 +23,11 @@
     "lodash": "^4.17.4"
   },
   "devDependencies": {
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.0.0",
+    "babel-preset-es2015": "^6.24.1",
     "eslint": "^3.16.0",
-    "pre-commit": "^1.2.2"
+    "pre-commit": "^1.2.2",
+    "webpack": "^2.4.1"
   }
 }
